requirements.txt
Flask
flask-cors
numpy
pandas
wfdb
neurokit2
PyWavelets
scipy
matplotlib
scikit-learn
imbalanced-learn
optuna
gunicorn
Flask-JWT-Extended
flask_migrate
Flask-Migrate
psycopg2-binary
Flask-SQLAlchemy
psycopg2-binary
Flask-Migrate
python-dotenv
supabase
pyedflib
python-json-logger
storage3
marshmallow